Initialization + Homepage and License
Initialized new Rails project. Styled the header and footer using CSS,
and added a link to the new license page in the footer. The header is
sticky, which means it stays at the top of the page when you scroll
(this uses some very basic JavaScript). The homepage just has one
sentence on it for now, "Welcome to PotionHW". The license page simply
reads the GNU AGPLv3 from LICENSE.txt and displays it, with blank lines
being converted to a new paragraph tag.